+++ src/GAME5_2.c
@@ -8677,4 +8677,3 @@ int nox_xxx_testCPUID2_444D90() { return 0; }
 #endif // NOX_CGO
 void nox_xxx_j_allocHitArray_511840(void) { nox_xxx_allocHitArray_5486D0(); }
 
-// ALL OK, 8062 function(s) have been successfully decompiled