+++ src/server__script__panic.h
@@ -3,9 +3,9 @@
 
 #include <stdbool.h>
 
-bool nox_script_panic_memhack_call(int fi, int* ret);
-bool nox_script_panic_memhack_check(int fnc);
-int nox_script_panic_memhack_sub_751090();
-int nox_script_panic_memhack_write_7510AC();
+bool nox_script_panic_compiler_call(int fi, int* ret);
+bool nox_script_panic_compiler_check(int fnc);
+int nox_script_panic_compiler_sub_751090();
+int nox_script_panic_compiler_write_7510AC();
 
 #endif // NOX_SERVER_SCRIPT_PANIC_H